Improve code
Use packages appropriately
Create Asset interface and use it in animal
Update tests
