--- src/main/java/Animal.java
@@ -1,6 +1,0 @@
-public class Animal {
-
-    public String speak() {
-        return null;
-    }
-}
