--- src/main/java/Bird.java
@@ -1,5 +1,0 @@
-public class Bird extends Animal {
-    public void fly() {
-        System.out.println("Whee ...");
-    }
-}
