+++ src/main/java/circus/Asset.java
@@ -1,0 +1,5 @@
+package circus;
+
+public interface Asset {
+    int getValue();
+}
