+++ src/main/java/circus/animal/Animal.java
@@ -1,0 +1,8 @@
+package circus.animal;
+
+import circus.Asset;
+
+public abstract class Animal implements Asset {
+
+    public abstract String speak();
+}
