+++ src/main/java/circus/animal/Duck.java
@@ -1,7 +1,14 @@
+package circus.animal;
+
 public class Duck extends Bird {
     @Override
     public String speak() {
         return "Quack Quack";
+    }
+
+    @Override
+    public int getValue() {
+        return 10;
     }
 
     @Override
