+++ src/main/java/circus/animal/Parrot.java
@@ -1,7 +1,14 @@
+package circus.animal;
+
 public class Parrot extends Bird {
     @Override
     public String speak() {
         return "Polly wants a cracker";
+    }
+
+    @Override
+    public int getValue() {
+        return 20;
     }
 
     @Override
