+++ src/main/java/circus/stuff/Cannon.java
@@ -1,3 +1,5 @@
+package circus.stuff;
+
 public class Cannon extends Equipment {
 
     public Cannon(int purchasePrice) {
