+++ src/main/java/circus/stuff/Equipment.java
@@ -1,11 +1,11 @@
-public abstract class Equipment {
+package circus.stuff;
+
+import circus.Asset;
+
+public abstract class Equipment implements Asset {
     protected int purchasePrice;
 
     public Equipment(int purchasePrice) {
         this.purchasePrice = purchasePrice;
-    }
-
-    public int getValue() {
-        return purchasePrice;
     }
 }
