+++ src/main/java/circus/stuff/Ladder.java
@@ -1,4 +1,8 @@
-public class Ladder extends Equipment {
+package circus.stuff;
+
+import circus.Asset;
+
+public class Ladder extends Equipment implements Asset {
 
     public Ladder(int purchasePrice) {
         super(purchasePrice);
